Refactored Settings using the new VocableCollectionViewController
